app/requirements.txt
asgiref==3.7.2
audioread==3.0.1
base58==2.1.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
Django==3.2
django-appconf==1.0.5
django-environ==0.11.2
django-filter==23.5
django-multiselectfield==0.1.12
django-oauth-toolkit==2.3.0
django-select2==8.2.1
django-video-encoding==1.0.0
djangorestframework==3.14.0
drf-spectacular==0.28.0
et-xmlfile==1.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
jdcal==1.4.1
joblib==1.3.2
jsonschema>=4.0.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.0
msgpack==1.0.7
mutagen==1.47.0
numba==0.58.0
numpy==1.25.2
openpyxl==3.0.10
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
psycopg2==2.9.9
pycparser==2.21
pytz==2023.3.post1
pyyaml==6.0.2
requests==2.31.0
resampy==0.4.2
scikit-learn==1.3.1
scipy==1.11.3
shutilwhich==1.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.6
